skeleton code for old PFI migration

+++ common/migrations/db/m180609_055058_new_pfi_schema_changes_.php
@@ -29,11 +29,10 @@ class m180609_055058_new_pfi_schema_changes_ extends Migration
     // Use up()/down() to run migration code without a transaction.
     public function up()
     {
-        $this->dropTable('proforma_invoice');
-        $this->dropTable('proforma_line_item');
-        $this->dropTable('proforma_item_lesson');
-        $this->dropTable('proforma_item_invoice');
         $tableSchema = Yii::$app->db->schema->getTableSchema('proforma_invoice');
+        if ($tableSchema) {
+            $this->dropTable('proforma_invoice');
+        }
         if ($tableSchema == null) {
             $this->createTable('proforma_invoice', [
                 'id' => $this->primaryKey(),
@@ -43,6 +42,9 @@ class m180609_055058_new_pfi_schema_changes_ extends Migration
             ]);
         }
         $Pfi_Line_Item = Yii::$app->db->schema->getTableSchema('proforma_line_item');
+        if ($Pfi_Line_Item) {
+            $this->dropTable('proforma_line_item');
+        }
         if ($Pfi_Line_Item == null) {
             $this->createTable('proforma_line_item', [
                 'id' => $this->primaryKey(),
@@ -50,6 +52,9 @@ class m180609_055058_new_pfi_schema_changes_ extends Migration
             ]);
         }
         $proforma_item_lesson = Yii::$app->db->schema->getTableSchema('proforma_item_lesson');
+        if ($proforma_item_lesson) {
+            $this->dropTable('proforma_item_lesson');
+        }
         if ($proforma_item_lesson == null) {
             $this->createTable('proforma_item_lesson', [
                 'id' => $this->primaryKey(),
@@ -58,6 +63,9 @@ class m180609_055058_new_pfi_schema_changes_ extends Migration
             ]);
         }
         $proforma_item_invoice = Yii::$app->db->schema->getTableSchema('proforma_item_invoice');
+        if ($proforma_item_invoice) {
+            $this->dropTable('proforma_item_invoice');
+        }
         if ($proforma_item_lesson == null) {
             $this->createTable('proforma_item_invoice', [
                 'id' => $this->primaryKey(),
